replace all the inputs with the custom input component

+++ src/components/Unsubscribe.vue
@@ -2,9 +2,8 @@
   <div>
     <h1>Désinscription à la newsletter</h1>
     <div class="form-group">
-      <label for="email">Email à désinscrire</label>
-      <input type="email" class="form-control" id="email" placeholder="nom@domain.ch" v-model.lazy="email">
-      <button class="btn btn-primary" @click="unsubscribe" style="margin-top:10px;">Confirmer</button>
+      <custom-input ref="email" label="Email à désinscrire" type="text" v-model="email" placeholder="nom@domain.ch" min="1" regexp="^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$" errorMessage="L'email n'est pas valide."></custom-input>
+      <button class="btn btn-primary" @click="unsubscribe">Confirmer</button>
       <div id="success-alert" class="alert alert-success" role="alert">
         <i class="fa fa-check" aria-hidden="true"></i> Désinscrit!
       </div>
@@ -22,7 +21,7 @@ import api from '../../config/api.js'
 export default {
   data () {
     return {
-      email: null
+      email: ''
     }
   },
   mounted: function () {
@@ -30,6 +29,8 @@ export default {
   },
   methods: {
     unsubscribe () {
+      const isEmailValid = this.$refs.email.isValid(this.email)
+      if (!isEmailValid) return
       this.$http.delete(`${api.subscriptions}/${this.email}`)
       .then((response) => {
         alert.show('#success-alert')
